Changed code to avoid incompatability with old perl on the QA server
Uncertain why we didn't run into this in the previous testing but a
limitation has been identified in the perl on the QA server.

The code:
      $pet_element = "\"p_id\":\"" . shift(keys(%patient_ids)) . "\"";
gets an error because shift will not accept keys as a parameter.  This
was replaced by the lines below which extract the keys in one step into
an array and then use the array in the shift.

      my @patients = keys(%patient_ids);
      $pet_element = "\"p_id\":\"" . shift(@patients) . "\"";